# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/trips": {
      "post": {
        "operationId": "TripsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar uma nova viagem",
        "tags": [
          "Trips - App"
        ]
      },
      "get": {
        "operationId": "TripsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar as viagens do usuário logado",
        "tags": [
          "Trips - App"
        ]
      }
    },
    "/trips/{id}": {
      "get": {
        "operationId": "TripsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhes de uma viagem específica",
        "tags": [
          "Trips - App"
        ]
      },
      "patch": {
        "operationId": "TripsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar uma viagem",
        "tags": [
          "Trips - App"
        ]
      },
      "delete": {
        "operationId": "TripsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletar uma viagem",
        "tags": [
          "Trips - App"
        ]
      }
    },
    "/trips/{id}/days": {
      "post": {
        "operationId": "TripsController_createDay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripDayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adicionar um dia no roteiro da viagem",
        "tags": [
          "Trips - App"
        ]
      }
    },
    "/trip-days/{id}": {
      "patch": {
        "operationId": "TripDaysController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripDayDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar dados de um dia específico da viagem",
        "tags": [
          "Trip Days - App"
        ]
      },
      "delete": {
        "operationId": "TripDaysController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletar um dia da viagem",
        "tags": [
          "Trip Days - App"
        ]
      }
    },
    "/trip-days/{id}/items": {
      "post": {
        "operationId": "TripDaysController_createItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItineraryItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adicionar um novo item/atividade ao dia da viagem",
        "tags": [
          "Trip Days - App"
        ]
      }
    },
...
    "/admin/users/{id}/travel-profile": {
      "patch": {
        "operationId": "AdminController_updateTravelProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserTravelProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar perfil de viagem de um usuário (Admin)",
        "tags": [
          "Admin - Users"
        ]
      },
      "get": {
        "operationId": "UserTravelProfileController_getAdminUserTravelProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Visualizar perfil do viajante de um usuário específico (Admin)",
        "tags": [
          "Travel Profile - App",
          "Admin - Users"
        ]
      }
    },
...
    "/admin/trips/{id}": {
      "get": {
        "operationId": "AdminTripsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalhar viagem completa",
        "tags": [
          "Admin - Trips",
          "Admin - Trips"
        ]
      },
      "patch": {
        "operationId": "AdminTripsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Editar viagem como ADMIN",
        "tags": [
          "Admin - Trips",
          "Admin - Trips"
        ]
      },
      "delete": {
        "operationId": "AdminTripsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Excluir viagem como ADMIN",
        "tags": [
          "Admin - Trips",
          "Admin - Trips"
        ]
      }
    },
...
    "/admin/trips/{tripId}/days": {
      "post": {
        "operationId": "AdminTripsController_createDay",
        "parameters": [
          {
            "name": "tripId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripDayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar dia em uma viagem",
        "tags": [
          "Admin - Trips",
          "Admin - Itinerary Editor"
        ]
      }
    },
    "/admin/trip-days/{id}": {
      "patch": {
        "operationId": "AdminTripsController_updateDay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripDayDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Editar dia de viagem",
        "tags": [
          "Admin - Trips",
          "Admin - Itinerary Editor"
        ]
      },
      "delete": {
        "operationId": "AdminTripsController_removeDay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Excluir dia de viagem",
        "tags": [
          "Admin - Trips",
          "Admin - Itinerary Editor"
        ]
      }
    },
    "/admin/trip-days/{tripDayId}/items": {
      "post": {
        "operationId": "AdminTripsController_createItineraryItem",
        "parameters": [
          {
            "name": "tripDayId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItineraryItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar item no roteiro",
        "tags": [
          "Admin - Trips",
          "Admin - Itinerary Editor"
        ]
      }
    },
...
    "/users/me/travel-profile": {
      "post": {
        "operationId": "UserTravelProfileController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserTravelProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar perfil do viajante",
        "tags": [
          "Travel Profile - App",
          "Travel Profile - App"
        ]
      },
      "get": {
        "operationId": "UserTravelProfileController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Visualizar perfil do viajante",
        "tags": [
          "Travel Profile - App",
          "Travel Profile - App"
        ]
      },
      "patch": {
        "operationId": "UserTravelProfileController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserTravelProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar perfil do viajante",
        "tags": [
          "Travel Profile - App",
          "Travel Profile - App"
        ]
      },
      "delete": {
        "operationId": "UserTravelProfileController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover perfil do viajante",
        "tags": [
          "Travel Profile - App",
          "Travel Profile - App"
        ]
      }
    },
...
  },
  "info": {
    "title": "2GO / AppRoteiros API",
    "description": "API de Roteiros de Viagem com IA e Autenticação OTP",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTripDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Viagem para Paris"
          },
          "destination": {
            "type": "string",
            "example": "Paris, França"
          },
          "coverImage": {
            "type": "string",
            "example": "https://exemplo.com/imagem.jpg"
          },
          "startDate": {
            "type": "string",
            "example": "2027-01-10T00:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "example": "2027-01-20T00:00:00.000Z"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "COMPLETED"
            ],
            "default": "DRAFT"
          },
          "preferences": {
            "type": "object",
            "example": {
              "focus": "cultural",
              "pace": "relaxed"
            }
          }
        },
        "required": [
          "title",
          "destination"
        ]
      },
      "UpdateTripDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Viagem para Paris"
          },
          "destination": {
            "type": "string",
            "example": "Paris, França"
          },
          "coverImage": {
            "type": "string",
            "example": "https://exemplo.com/imagem.jpg"
          },
          "startDate": {
            "type": "string",
            "example": "2027-01-10T00:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "example": "2027-01-20T00:00:00.000Z"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "COMPLETED"
            ],
            "default": "DRAFT"
          },
          "preferences": {
            "type": "object",
            "example": {
              "focus": "cultural",
              "pace": "relaxed"
            }
          }
        }
      },
      "CreateTripDayDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2027-01-10T00:00:00.000Z"
          },
          "dayNumber": {
            "type": "number",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Chegada e Check-in"
          },
          "description": {
            "type": "string",
            "example": "Dia livre para explorar os arredores do hotel."
          }
        },
        "required": [
          "dayNumber"
        ]
      },
      "UpdateTripDayDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2027-01-10T00:00:00.000Z"
          },
          "dayNumber": {
            "type": "number",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Chegada e Check-in"
          },
          "description": {
            "type": "string",
            "example": "Dia livre para explorar os arredores do hotel."
          }
        }
      },
      "CreateItineraryItemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Visita ao Museu do Louvre"
          },
          "description": {
            "type": "string",
            "example": "Comprar ingressos antecipados."
          },
          "category": {
            "type": "string",
            "enum": [
              "TOURIST_ATTRACTION",
              "MUSEUM",
              "RESTAURANT",
              "CAFE",
              "BAR",
              "BEACH",
              "PARK",
              "SHOPPING",
              "EXPERIENCE",
              "TRANSPORT",
              "EVENT",
              "NIGHTLIFE",
              "FREE_ACTIVITY",
              "PAID_ACTIVITY"
            ]
          },
          "location": {
            "type": "string",
            "example": "Rue de Rivoli, 75001 Paris"
          },
          "googleMapsLink": {
            "type": "string",
            "example": "https://maps.google.com/?q=Louvre"
          },
          "latitude": {
            "type": "number",
            "example": 48.8606111
          },
          "longitude": {
            "type": "number",
            "example": 2.337644
          },
          "timeLabel": {
            "type": "string",
            "example": "09:00"
          },
          "period": {
            "type": "string",
            "example": "Manhã"
          },
          "duration": {
            "type": "number",
            "example": 120,
            "description": "Duração em minutos"
          },
          "cost": {
            "type": "number",
            "example": 15.5
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          },
          "externalLink": {
            "type": "string",
            "example": "https://louvre.fr"
          },
          "notes": {
            "type": "string",
            "example": "Levar garrafa de água."
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "isEditable": {
            "type": "boolean",
            "default": true
          },
          "isUserModified": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "title",
          "category",
          "order"
        ]
      },
...
      "UpdateUserTravelProfileDto": {
        "type": "object",
        "properties": {
          "bio": {
            "type": "string"
          },
          "preferredStyles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ECONOMIC",
                "COMFORT",
                "LUXURY",
                "ADVENTURE",
                "FAMILY",
                "ROMANTIC",
                "PARTY",
                "CULTURAL"
              ]
            }
          },
          "budgetLevel": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "PREMIUM"
            ]
          },
          "favoriteCountries": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "favoriteCities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredLanguages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "foodPreferences": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "accessibilityNeeds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "travelInterests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "travelCompanions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "avoidedDestinations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredClimate": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bucketListDestinations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "prefersNightlife": {
            "type": "boolean"
          },
          "prefersNature": {
            "type": "boolean"
          },
          "prefersGastronomy": {
            "type": "boolean"
          },
          "prefersMuseums": {
            "type": "boolean"
          },
          "prefersShopping": {
            "type": "boolean"
          },
          "prefersRelaxing": {
            "type": "boolean"
          },
          "averageTripDuration": {
            "type": "number"
          },
          "passportCountry": {
            "type": "string"
          },
          "instagramHandle": {
            "type": "string"
          }
        }
      },
...
      "CreateUserTravelProfileDto": {
        "type": "object",
        "properties": {
          "bio": {
            "type": "string"
          },
          "preferredStyles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ECONOMIC",
                "COMFORT",
                "LUXURY",
                "ADVENTURE",
                "FAMILY",
                "ROMANTIC",
                "PARTY",
                "CULTURAL"
              ]
            }
          },
          "budgetLevel": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "PREMIUM"
            ]
          },
          "favoriteCountries": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "favoriteCities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredLanguages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "foodPreferences": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "accessibilityNeeds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "travelInterests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "travelCompanions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "avoidedDestinations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preferredClimate": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bucketListDestinations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "prefersNightlife": {
            "type": "boolean"
          },
          "prefersNature": {
            "type": "boolean"
          },
          "prefersGastronomy": {
            "type": "boolean"
          },
          "prefersMuseums": {
            "type": "boolean"
          },
          "prefersShopping": {
            "type": "boolean"
          },
          "prefersRelaxing": {
            "type": "boolean"
          },
          "averageTripDuration": {
            "type": "number"
          },
          "passportCountry": {
            "type": "string"
          },
          "instagramHandle": {
            "type": "string"
          }
        }
      },
...
    }
  }
}